+++ atoman/filtering/filters/clusterFilter.py
@@ -35,7 +35,20 @@ Parameters are:
     
     Hide atoms
         If `Draw convex hulls` is selected this will make the atoms invisible,
-        so just the hulls are shown.
+        so just the hulls are shown. Cannot be selected at the same time as
+        `Show atoms inside hulls` or `Show atoms outside hulls`.
+    
+    Show atoms inside hulls
+        If `Draw convex hulls` is selected this will make all atoms that fall
+        within a convex hull visible, regardless of previous filters (i.e. it
+        acts on the original input lattice). Cannot be selected at the same
+        time as `Hide atoms` or `Show atoms outside hulls`.
+    
+    Show atoms outside hulls
+        If `Draw convex hulls` is selected this will make all atoms that fall
+        outside the convex hulls visible, regardless of previous filters (i.e. it
+        acts on the original input lattice). Cannot be selected at the same
+        time as `Hide atoms` or `Show atoms inside hulls`.
     
     Calculate volumes
         Calculate the volumes of the clusters of atoms.
@@ -52,7 +65,11 @@ Parameters are:
 """
 from __future__ import absolute_import
 from __future__ import unicode_literals
+from __future__ import division
+import copy
+
 import numpy as np
+from scipy.spatial import Delaunay
 
 from . import base
 from .. import clusters
@@ -72,6 +89,8 @@ class ClusterFilterSettings(base.BaseSettings):
         self.registerSetting("calculateVolumesVoro", default=True)
         self.registerSetting("calculateVolumesHull", default=False)
         self.registerSetting("hideAtoms", default=False)
+        self.registerSetting("showAtomsInHulls", default=False)
+        self.registerSetting("showAtomsOutHulls", default=False)
         self.registerSetting("neighbourRadius", default=5.0)
         self.registerSetting("hullCol", default=[0, 0, 1])
         self.registerSetting("hullOpacity", default=0.5)
@@ -140,6 +159,56 @@ class ClusterFilter(base.BaseFilter):
             clusterListIndex = clusterIndexMapper[clusterIndex]
             clusterList[clusterListIndex].addAtom(atomIndex)
         
+        # show all atoms inside or outside the clusters
+        drawHulls = settings.getSetting("drawConvexHulls")
+        showInHulls = settings.getSetting("showAtomsInHulls")
+        showOutHulls = settings.getSetting("showAtomsOutHulls")
+        if drawHulls and (showInHulls or showOutHulls):
+            # first we calculate the Delaunay triangulation of each cluster (including periodic images)
+            self.logger.debug("Calculating Delaunay triangulations for showing atoms")
+            hulls, hullsMap = self.computeDelaunayForClusters(lattice, clusterList, nebRad)
+            
+            # for each atom determine whether it lies within a cluster or not
+            self.logger.debug("Determining location of atoms (inside or outside clusters)")
+            # TODO: write in C
+            inClusterMask = np.zeros(lattice.NAtoms, np.int32)
+            pos = np.empty((1, 3), np.float64)
+            for i in range(lattice.NAtoms):
+                pos[0][:] = lattice.atomPos(i)[:]
+                for hull, hullMap in zip(hulls, hullsMap):
+                    res = hull.find_simplex(pos) >= 0
+                    if res[0]:
+                        # set the mask to in cluster
+                        inClusterMask[i] = 1
+                        
+                        # add to the cluster if doesn't already belong
+                        cluster = clusterList[hullMap]
+                        if i not in cluster:
+                            cluster.addAtom(i)
+                        
+                        break
+            
+            # make sure all cluster atoms are included
+            for cluster in clusterList:
+                for index in cluster:
+                    inClusterMask[index] = 1
+            
+            # make the new visible atoms array, starting with full system
+            self.logger.info("Overriding visible atoms based on cluster occupancy")
+            visibleAtoms.resize(lattice.NAtoms, refcheck=False)
+            visibleMask = 1 if showInHulls else 0
+            # TODO: write in C
+            numVisible = 0
+            for i in range(lattice.NAtoms):
+                if inClusterMask[i] == visibleMask:
+                    visibleAtoms[numVisible] = i
+                    numVisible += 1
+            visibleAtoms.resize(numVisible, refcheck=False)
+            
+            # TODO: set cluster list to be empty on case of show out
+            if showOutHulls:
+                clusterList = []
+        
         # calculate volumes
         if calcVols:
             self.logger.debug("Calculating cluster volumes")
@@ -153,7 +222,7 @@ class ClusterFilter(base.BaseFilter):
                     self.logger.debug("Cluster %d: facet area is %f", i, area)
         
         # hide atoms if required
-        if settings.getSetting("drawConvexHulls") and settings.getSetting("hideAtoms"):
+        if drawHulls and settings.getSetting("hideAtoms"):
             visibleAtoms.resize(0, refcheck=False)
         
         # result
@@ -161,3 +230,46 @@ class ClusterFilter(base.BaseFilter):
         result.setClusterList(clusterList)
         
         return result
+    
+    def computeDelaunayForClusters(self, lattice, clusterList, neighbourRadius):
+        """Compute Delaunay triangulation for each cluster, including periodic images."""
+        # build and store convex hulls for each cluster (unapply PBCs!?)
+        self.logger.debug("Computing Delaunay for each hull (unapplying PBCs)")
+        cellDims = lattice.cellDims
+        hulls = []
+        hullClusterMap = []
+        for clusterIndex, cluster in enumerate(clusterList):
+            appliedPBCs = np.zeros(7, np.int32)
+            clusterPos = cluster.makeClusterPos()
+            
+            _clusters.prepareClusterToDrawHulls(len(cluster), clusterPos, cellDims, np.ones(3, np.int32), appliedPBCs,
+                                                neighbourRadius)
+            
+            hulls.append(self.makeDelaunay(clusterPos))
+            hullClusterMap.append(clusterIndex)
+            
+            # handle PBCs here
+            while max(appliedPBCs) > 0:
+                tmpClusterPos = copy.deepcopy(clusterPos)
+                clusters.applyPBCsToCluster(tmpClusterPos, cellDims, appliedPBCs)
+                hulls.append(self.makeDelaunay(tmpClusterPos))
+                hullClusterMap.append(clusterIndex)
+        
+        return hulls, hullClusterMap
+    
+    def makeDelaunay(self, clusterPos):
+        """Calculate Delaunay for the given position."""
+        # make pts
+        # TODO: C or view
+        num = len(clusterPos) // 3
+        pts = np.empty((num, 3), np.float64)
+        for i in range(num):
+            i3 = 3 * i
+            pts[i][0] = clusterPos[i3]
+            pts[i][1] = clusterPos[i3 + 1]
+            pts[i][2] = clusterPos[i3 + 2]
+        
+        # make hull
+        hull = Delaunay(pts)
+        
+        return hull