+++ atoman/gui/mainWindow.py
@@ -637,6 +637,7 @@ class MainWindow(QtWidgets.QMainWindow):
         """
         shutil.rmtree(self.tmpDirectory)
         self.console.accept()
+        self.threadPool.waitForDone()
     
     def hideProgressBar(self):
         """