+++ atoman/gui/outputDialog.py
@@ -682,17 +682,15 @@ class RDFForm(genericForm.GenericForm):
             return
         
         # then determine species
-        specieList = inputLattice.specieList
-        
         if self.spec1 == "ALL":
             spec1Index = -1
         else:
-            spec1Index = int(np.where(specieList == self.spec1)[0][0])
+            spec1Index = inputLattice.getSpecieIndex(self.spec1)
         
         if self.spec2 == "ALL":
             spec2Index = -1
         else:
-            spec2Index = int(np.where(specieList == self.spec2)[0][0])
+            spec2Index = inputLattice.getSpecieIndex(self.spec2)
         
         # rdf calulator
         rdfCalculator = rdf.RDFCalculator()
@@ -924,11 +922,11 @@ class ImageTab(QtWidgets.QWidget):
         renderTypeButtonGroup.setExclusive(1)
         renderTypeButtonGroup.buttonClicked[int].connect(self.setRenderType)
         
-        self.POVButton = QtWidgets.QPushButton(QtGui.QIcon(iconPath("other/pov-icon.svg")), "POV-Ray")
+        self.POVButton = QtWidgets.QPushButton(QtGui.QIcon(iconPath("other/pov-icon.png")), "POV-Ray")
         self.POVButton.setCheckable(1)
         self.POVButton.setChecked(0)
         
-        self.VTKButton = QtWidgets.QPushButton(QtGui.QIcon(iconPath("other/vtk-icon.svg")), "VTK")
+        self.VTKButton = QtWidgets.QPushButton(QtGui.QIcon(iconPath("other/vtk-icon.png")), "VTK")
         self.VTKButton.setCheckable(1)
         self.VTKButton.setChecked(1)